P2 FINALLY WORKS!!!
It was a long, arduous and annoying process. Not even an impactful one too but P2's animator finally isn't broken. I just had to change it from the override animator to a normal animator and it FINALLY WORKS. It sucks that I'm still in the same spot as before, I've still got so many assets to draw, add and script, but this is an achievement in my books!