Health Bars are... functional.
The added health bars do work but the way the Udemy tutorial goes about doing it is confusing. I'm really scared of scrapping and remaking assets but the fact that the Udemy's tutorial is following a 3D fighting game is probably to blame. I need to look at other tutorials for how to implement combat because it has not been working out.