Health Bars are up and running!
Health weren't that hard to set up, but it's funny how after what feels like 6 months or so, I'm finally ahead of where my prototype was. There's still a lot of work to do, such as getting the rest of the UI, the other screens and most importantly, the other characters. But it only goes up from here!