I found the reason why, now I just don't know how...
For whatever reason, when the Player1Layer0 is "Blocking" the colliders it's supposed to be disabling isn't disabled. If I change the Tag to any other state like crouching, the code works as intended. So I figured that out, I just don't know why and how to fix it. In the meantime, I'm gonna fix P2's animations because it bugs everytime the players swaps sides.