The major problem, and some minor problems.
To anyone who wants to access this project, there are two problems I'm trying to fix. One major and one minor. More details in the README.md

The Main problem: The 'Standing Block' doesn't works as intended. Instead of disabling the colliders that would calculate damage, the colliders flicker between on and off, therefore taking more damage for as long as an attack is active. Doesn't seem to be a problem with the code because the crouching block follows the same thing and works perfectly, so it might have something to do with the 'Standing Block' animation itself or it's transitions. I don't know, which is why I'm asking for help.

The minor problem: The damage animation that plays when your character is hit, doesn't play when they're facing left (i.e. on the right side of the screen). The Animator says that the animations are transitioning just fine so I don't know what the problem is.