Experimenting with Knockback: Precursor
The blocking problem is still present but I'm moving forward still until I can figure out the cause. Making a save here because I want to experiment with player knockback, so if the player is hit on the ground or in the air, they got knocked back. Doing this just in case something screws up.