Knockback is functional, and somehow it fixed the minor issue!?
The knockback works great! Just needs some fine-tuning to get it right. Funnily enough, I don't know how but now the damage animation doesn't cut short when the players are facing left so that's a win, even if it looks a little finicky! The blocking is still a problem though, need to find a solution on that.