Implementing AI
Blocking still doesn't work. I need to remind myself this every time I make an update. I also need to remind myself that for the characters to spawn properly, the prefab needs to be dragged out from the prefab folder than from the hierarchy, else chaos ensues. With that said, I'm beginning to implement an AI character for the player to fight. The problem is that it requires a new set of code so the AI can move, which I fear could be a problem but I'm just gonna have to trust the process.