AI can move and attack
The AI can do the bare minimum for being an AI. They can move towards the character and throw attacks at random, though it's just the one attack for some reason. Also note, be careful not to jump directly on top of the AI because the way I've coded it, the AI will always track the player's position across the x and y axis' instead of a proximity box (that's just the the udemy tutorial did it) so if you land on the head, you're stuck there. Other than that, it's functioning.