Small back-up for safety reasons
Not much has changed, just made it so the attack rate can be manually adjusted in the inspector. I've also increased the knock back for the heavy attack. Planning to give my AI the ability to block. Then once I've finished some more animations, and could try to give my character a projectile to fire.