Main Menu & Other Menus
Okay, so it's been a good while since I've made a back-up (because I kept on forgetting) but I've made quite a few advancements. That is, I've got a functioning main menu! I've rearranged the orders of the scenes now that I have 6 of them. The most important part about the main menu is that you can choose between One & Two player mode now! I've been fooling around with the particle system as well, so the main menu screen and the character select screen has a little flair. I'm working on make the Options menu so the player can adjust some settings like volume and brightness. And of course, the biggest challenge is making the characters, background and music in time. But I feel confident I can do this!