I made a build. I MADE A GAME!!
Ok so it's been a LONG WHILE since I made a backup so there's a ton to unpack here. Most important detail though, this is the version that I sending in for my Major Work. This is THE Version 0.1 of Flora! I'll keep it brief and say that I've added sfx and voices, I've implemented Gilford but just his idle animation, I've created AI of Bri and Gil, added another volume slider in the options and finally there's a credit screen. It sounds short but I did a shit ton. And thing is, I'd say this project is only 70% done. I'm gonna keep working on Flora as I grow as a game dev. But for now, I can confidently say that I've created a full fledged video game!